Merge internal changes into public repository (change 246251381)

+++ models/official/mask_rcnn/spatial_transform_ops.py
@@ -156,13 +156,13 @@ def selective_crop_and_resize(features,
       tf.reshape(x_indices, [batch_size, num_boxes, 1, output_size * 2]) *
       tf.ones([1, 1, output_size * 2, 1], dtype=tf.int32))
 
-  # Temporary variables with int32 dtype is not supported for GPU's. To make
-  # model_fn implementation be GPU friendly, we add casting to int64 here.
+  # TODO(hongjunchoi): Remove the need for temporary variables as
+  # temporary variables with int32 dtype are not supported for GPU's.
   indices = tf.add_n([
-      tf.cast(batch_dim_indices, tf.int64),
-      tf.cast(box_level_indices, tf.int64),
-      tf.cast(height_indices, tf.int64),
-      tf.cast(width_indices, tf.int64)
+      batch_dim_indices,
+      box_level_indices,
+      height_indices,
+      width_indices,
   ])
   indices = tf.reshape(indices, [-1])
